+++ tests/__init__.py
@@ -2,4 +2,6 @@ import logging
 
 import bugsnag
 
+
+logging.getLogger('requests').setLevel(logging.CRITICAL)
 bugsnag.logger.setLevel(logging.CRITICAL)